Shrink `SyncState`
std may need to support 2^64 threads, but we will be very fine with 2^15-1. I even implemented a machinery to reuse thread IDs, to be extra safe.

After implementing this I realized we can just use a thread_local with a global atomic, just like std, and this will probably simplify the implementation, but this won't allow us to reuse IDs.